MB-67053 : Add additional checks in the /admin/log endpoint handler
The /admin/log endpoint returns the contents of a user-specified file.

When a user specifies a file, the handler generates its file path by
joining the file name to the log directory path.

Add a check to verify that the file path generated is not for a file
outside the log directory.

Change-Id: I4305747be3e23c96ffc8a678f1e33b7055e66f1b
Reviewed-on: https://review.couchbase.org/c/query/+/228786
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>