MB-67060 Add nil pointer check
The issue is similar to that for an earlier fix for MB-55597.
When going through join enumeration, if we decided to perform
bit-filter pushdown (generating a bloom filter from build side
of a hash table and push to the probe side, to reduce fetch
size on the probe side), in certain cases we need to maintain
the bit-filter information, e.g. remove previously constructed
bit filters if it is no longer relevant. This involves setting
a pointer in a slice to nil, and subsequently compact the slice
to remove all nil entries. The issue is that if there are
multiple sender (for probe filter) or receiver (for build filter)
for the same slice, we would be walking the slice mulitple
times before we perform the compact operation. In such cases
it is possible to have one walk through the slice setting one
entry to nil and a second walk through the slice hit that
entry. The previous (part of the) fix for MB-55597 handled this
for the build filters, but not the probe filters which this
current bug now hits.

The fix is to apply the same fix (nil-check) for probe fitlers
as well.

Change-Id: Ida6bc4bc6175a2560b87af031e7852ddf15a5805
Reviewed-on: https://review.couchbase.org/c/query/+/228822
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>