MB-67053: [trinity] Add additional checks in the /admin/log...
endpoint handler

The /admin/log endpoint returns the contents of a user-specified file.

When a user specifies a file, the handler generates its file path by
joining the file name to the log directory path.

Add a check to verify that the file path generated is not for a file
outside the log directory.

Change-Id: Iac959f816c42e0ad615fa7cfb79f1278f7db0d8f
Reviewed-on: https://review.couchbase.org/c/query/+/228856
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>