MB-67089: [BP MB 64409] Fix error when parsing inline function's...
...expression string

The issue is due to the expression.UserDefinedFunction.String() method
returning the full name of the UDF function without backticks.

Example:
CREATE FUNCTION test1() { `func-1`()};
The execution of such a function would return a parsing error.

Issues caused by this:

1. When executing an inline UDF, the string of the body's expression is
parsed.

If the body contains a reference to a function whose full name contains
parts that require being escaped by backticks, the parsing fails.

2.  The unmarshalling of prepared statements that refer to such UDFs
would fail due parsing failure of the UDF's expression string.

And hence would only execute on the node the prepared statement
was created in.

Fix:
The fix is to return the full name of the UDF function with backticks
when calling the expression.UserDefinedFunction.String() method.

Change-Id: I9a85cbb86d76649b4dcc3fe8baef6f304747375c
Reviewed-on: https://review.couchbase.org/c/query/+/228908
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>