MB-66994: Fix spill file access errors when Order operator stopped
When the Ordering operators receive a stop signal, the spill files
associated with the sort array are closed.

But, the Ordering operator's execution in another routine can be
accessing the sort array and its spill files conccurrently.

This can result in "file closed" type errors being returned to the user
which is misleading.

To prevent this, ignore any errors retured when accessing spill files
if the operator is stopped.

Change-Id: I69f01bdb60996727b5eb066598146ed61a843581
Reviewed-on: https://review.couchbase.org/c/query/+/229259
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>