MB-67241 Mark identifier expression properly when formalizing ...    ... correlated subquery in parser
When an embedded subquery text from a prepared statement is being
parsed on a secondary node, if the subquery is correlated, we go
through formalization and treat any identifiers not recognized as
a correlation reference. This is done by adding the identifier to
the correlation map. However, the identifier expression itself may
not be properly marked in this case, subsequently when covering
determination is done, we may incorrectly determine that covering
cannot be done due to lack of proper identifier flag.

The fix is to mark the identifier expression itself as both a
keyspace alias (which is implied if it is a correlated reference)
and correlated.

Change-Id: Iece32b9d9fa49fa5c6096515a5868fff6eecdd5a
Reviewed-on: https://review.couchbase.org/c/query/+/230070
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>