MB-67425 Remove check for PreserveProjection()...
   ...when executing a user defined function,
as inline udf's use subquery plans and
subquery plans are never built with project
operator preserving order of result terms,
therefore this check is not necessary.

By removing the check we reuse the old
context. As a result, when we have repeated
executions of an inline udf in request
the subquery evaluation can return the
cached result, which speeds up the execution.

Also, we explicitly set the flag to preserve
projection order to false before executing a function
and restore it after the execution returns.

Change-Id: I8ce79be6fb3f0a7c1c5025be6cd95ce3307f7772
Reviewed-on: https://review.couchbase.org/c/query/+/230641
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Gaurav J <gaurav.jayaraj@couchbase.com>
Well-Formed: Restriction Checker