MB-67606 Unset early order if intersect scan is used
When CBO is turned on, if multiple indexes can satisfy a query,
CBO picks the best number of indexes to be used, either a
single index scan or intersect scan with multiple indexes. The
issue with this bug is that if CBO picks multiple indexes, we
did not do the unsetting of early order flags which we do in
case of RBO (we do not reduce the number of indexes in RBO).
This results in failure in resetting index order and later on
failure to build an ORDER operator.

The fix is to also check the number of indexes in case of CBO,
and if it is > 1, also perform the unsetting of early order
flags.

Change-Id: I9dd3c14c904688662459913611f2025c59f5db5e
Reviewed-on: https://review.couchbase.org/c/query/+/230656
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>