MB-67432 Track value with SetField for objectValue
For objectValue (and sliceValue) the Track() and Recycle() calls
used to be no-op. As part of the fix for MB-63420, we now do
Track()/Recycle() calls for subelements of objectValue/sliceValue
when Track()/Recycle() on objectValue/sliceValue is called. What
is missing is that for annotatedValue and ScopeValue, in addition
to Track()/Recycle() calls, we also have a Track() call when
SetField() is called, the rationale is that with SetField(), we
now have a new reference to the value as part of the field for
the annotatedValue/ScopeValue. This same logic is currently
missing for objectValue, as a result, when we call SetField() on
an objectValue, we did not properly increment the reference
count for the field value, resulting in the value being released
prematurely.

The fix is to also add a Track() call as part of SetField() call
for objectValue, similar to what is done for annotatedValue and
ScopeValue. Note for sliceValue this is not relevant since the
SetField() call on sliceValue returns a "NOT SETTABLE" error.

For the special case where a missing value is used for the
SetField() call, effectively unsetting a field, also add a
Recycle() call if a value is previously existing.

Also added logic in SetField() call for annotatedValue and
ScopeValue such that if the underlying Value is objectValue, do
not call Track() again since it is already done in objectValue.

Change-Id: Ic2637805de2cc420d2e5ef1d9f00144c12e125d9
Reviewed-on: https://review.couchbase.org/c/query/+/230787
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>