MB-67450 Add new infer options
1. array_sample_size
If a non-negative value is provided, we proceed to
trim array samples to this value if the length of array
sample is greater than the value. The array field
would include "sampleSize" indicating this option was in effect.

2. max_nesting_depth
For this functionality we now track the depth of
every field we have visit. Using this depth information,
if a non-negative "max_nesting_depth" value is
provided we stop visiting fields that are
at a nesting depth which is greater than or equal to
this value. The depth information is included
as "nestingDepth" in the field's information
with 0 indicating a top-level field.

Change-Id: I3576f7737f0d18b0c7c3d919b55dee2eb949acd4
Reviewed-on: https://review.couchbase.org/c/query/+/230955
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>