MB-66842 Protect struct members for concurrent access
A value can be copied multiple times during execution time, e.g. when
UNNEST or JOIN is done. When a copy is made, it is possible for a
child value to be shared among multiple copies, since for an
objectValue or a sliceValue, we simply copy the top-level struct and
use the same children fields/elements in the copy. If such copies
are accessed/modified by different threads, e.g. during initial
projection, it is possible to have timing-related issues for
simultaneous accessing/manipulating of the same value structure.

For a parsedValue, there is already a read/write lock that is used
to protect a couple of struct members (fields, elements), however
there are a couple of additional struct members (raw, parsed, which
represents the unparsed and parsed values, respectively) that also
needs synchronization. The issue with this bug is that when these
two members are modified (mostly likely in unwrap() function, where
the actual parsing is done and switching from raw to parsed), it is
possible for other access to these two members to have inconsistent
values (e.g. accidentally getting a nil pointer) which may cause
issues.

The fix is to also protect these two members with the lock, i.e.,
write lock when modified and read lock when accessed. Also use
local variables after accessing these members such that we don't
get inconsistent values.

Change-Id: Id06bd4d38ccf860422d89c8569c8d0d5bed1ad64
Reviewed-on: https://review.couchbase.org/c/query/+/231211
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>