MB-67709 Add Track call for setting META fields
The fix for MB-67432 adds a Track() call for SetFields() and a
Recycle() call for UnsetFields(). For this bug, when we get
info from active requests, we put the query plan as a META
field and call UnsetField() to remove it as a regular field.
There are a couple of issues here, one is that when UnsetField()
is called we call Recycle() on the field value, but we never
called Track(), which was supposed to be called when SetField()
is called, except in this case we used SetMetaField() instead;
the second issue is that the value in this case is a
marshalledValue, and there is a mismatch between the behavior
of Track() vs Recycle() for a marshalledValue. For Track(),
we just call Track() for the unwrapped value, but for Recycle(),
current code just throw away the marshalledValue. There is no
refCnt on marshalledValue and thus we cannot easily determine
whether there are existing references to the value. This
"blind" recycling of a marshalledValue can be problematic.

To fix this, there are two parts:

 - add Track() call to the value used when SetMetaField() is
   called; also add Track() call in other places where we set
   a value to be other parts of the AnnotatedValue, such as
   adding it as Value, or as Attachment, or as Projection.
 - modify the way Track()/Recycle() is handled for
   marshalledValue, this includes:
    - adds a refCnt to marshalledValue
    - when Track()/Recycle() is called on marshalledValue,
      first check whether the value has been unwrapped, if
      not, then update the refCnt, and only do the current
      recycle of marshalledValue if refCnt is 0; if the
      value has been unwrapped, then simply pass along the
      Track()/Recycle() call to the underlying value
    - during unwrap(), if refCnt > 0, "transfer" the refCnt
      to Track() calls to the newly created marshalled value.
    - the modified mechanism for Track()/Recycle() ensures
      we still keep the original intent of marshalledValue,
      i.e. do not do the marshalling until necessary, but
      also integrate with the general infrastructure for
      Track()/Recyle() for values

In addition, also changed the function ObjectRename to call
SetField() (on the new field name) before UnsetField() (on the
old field name) such that the underlying value has a Track()
call before a Recycle() call.

Change-Id: If604d831e1535d4ac3278513a1a0b9b4258ba966
Reviewed-on: https://review.couchbase.org/c/query/+/231272
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/231274
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/232200