MB-66643: [BP] Fix incorrect deletion of CBO docs associated with long...
scope/collection names

When a bucket is loaded, a cleanup routine scans the _system._query
collection and deletes documents associated with scopes and
collections that no longer exist.

The scope/collection names are identified from the document key
of the CBO document. And the names are checked if they still
exist in the cluster. And if they do not, the document is
deleted.

But, scope/collection names longer than 36 characters have
UUIDs in place of actual names in the CBO document key.

Since typically no scope/collection has a UUID as
its name, these documents were incorrectly identified as stale
and deleted.

The fix is:

If the key contains a UUID, fetch the CBO document and extract
the actual scope/collection name from the document contents.
Use this information to determine if the document should be
deleted or not.

Change-Id: I83e074e41fb4c92d171ae19b5ec959b96c1d06df
Reviewed-on: https://review.couchbase.org/c/query/+/231331
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>