MB-67657: Fix memory session cleanup and termination for...
statistics updates executed by AUS

* AUS creates a memory session for each statistics update when
memory or node quota is set.

* Previously, a final release of this memory session was not done.
Which could cause left-over memory usage to never be released to
the manager.

* Added a final release of the session when the update completes.

* AUS previously used a workaround to stop statistics update by
sending a signal to the value connection. Remove this.

* Now stop statistics updates by marking the context as inactive,
ike regular UPDATE STATISTICS statements executed outside AUS.
AUS marks the context as inactive when the task stops or an error
occurs

Change-Id: I09eff952e927d433ed5cf865847eb77c566af98e
Reviewed-on: https://review.couchbase.org/c/query/+/231387
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>