MB-67549 Do not throw away index that could use index aggregation
When the planner compare two indexes, it first try to pick one with
more sargable index keys. This logic is normally sound since it is
unlikely that the index with less number of sargable keys to have
better properties (e.g. for index pushdown), since index pushdown
typically requires the index spans to be exact, and the index with
less number of sargable keys is unlikely to be exact (since it is
missing a sargable span). The exception to this, and is happening
in this bug, is that the "missing" sargable key is present in the
index in another index key, i.e. there is some redundancy in the
index keys. One typical such usage is the index has a regular key,
e.g. c1, but also has a functional key on c1, like abs(c1), in
which case if the query has a predicate on abs(c1), it can be used
to sarg both c1 and abs(c1), which counts as 2 sargable keys, and
current logic will favor an index with (c1, abs(c1)) over an index
with just (abs(c1)), even though the latter could be used with
index aggregation.

To fix this, we now check and mark sargable indexes with redundant
index keys (one index key depends on another index key), and when
comparing two indexes, if one index has more sargable index keys,
but also has redundant index keys, then we further compare index
pushdown properties, and reset the favored index to the one with
better pushdown properties. If the pushdown properties are the
same (e.g. when pushdown property has not been determined), then
reset to not favoring one over the other, and let other criteria
kicks in.

Change-Id: I61ea9f4e2d6001ef6503219e7d68aa2784195ec8
Reviewed-on: https://review.couchbase.org/c/query/+/231474
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker