Combined test: Make index generation configurable...
Created an option to configure whether the test should generate
and create indexes.

If set to false, index key definitions
specified in config.json are not considered nor is ADVISE
run on the generated queries.

However, indexes are still generated to satisfy queries with
JOINs. This is because in certain cases JOINs can return
errors if suitable indexes are not present.

Implemented this option so that the test can be configured to
test sequential scans.

Modified the configuration file to include request parameters
`auto_execute` and `auto_prepare`.

Made these changes so that the test can be configured to create
anonymous prepared statements for the generated queries and
execute them.

Change-Id: Ided0b37d9b85a5e67b50c52efa1d58d5391bad0d
Reviewed-on: https://review.couchbase.org/c/query/+/231527
Reviewed-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>