MB-67882 Allow intersect scan with FTS indexes and vector indexes
If a query has vector search with GSI/Bhive indexes, and also references
to FTS indexes (e.g. using SEARCH() function), currently we return an
error if intersect scan is used with both a vector index and FTS index.

Relax the restriction such that we only return an error if the vector
index scan is also performing reranking (which is not expected if FTS
index scan is also used, in which case we won't have exact spans).

Additional logic is addes such that if an index (GSI or FTS) can be
used for vector search, and also provides ordering, then just use
that index, avoid intersect scans in this case.

A secondary issue uncovered is that if the vector index scan is using
early order, but later we add a FTS index scan, we need to unset the
early order. This logic is already done if we have multiple GSI
indexes used in an intersect scan, but is missed if we have an
intersect scan with a GSI index and a FTS index. (This part of the fix
should be backported to 7.6 code line).

Change-Id: Ic5d3cced3c020356972af9f2d9ad00661202496c
Reviewed-on: https://review.couchbase.org/c/query/+/231843
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>