MB-67947 Unset early order in case of intersect scan with FTS indexes
Logic is addes such that if a FTS index can be used for vector
search, and also provides ordering, then just use that index,
avoid intersect scans in this case.

In addition, if an index scan is using early order, but later we
add a FTS index scan, we need to unset the early order. This logic
is already done if we have multiple GSI indexes used in an
intersect scan, but is missed if we have an intersect scan with
a GSI index and a FTS index.

Change-Id: Ic5d3cced3c020356972af9f2d9ad00661202496c
Reviewed-on: https://review.couchbase.org/c/query/+/231843
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/231853
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/231871