MB-67882 Do not return early if SEARCH() has KNN
The previous fix allows a GSI index to be chosen by itself if the
index can do vector search and also has order pushdown. One rare
case that needs to be handled is that if the query also has a
SEARCH() function, and the SEARCH() function is also doing
vector search, then the SEARCH() function cannot be evaluated as
a filter, it must be evaluated by IndexFtsSearch (see MB-62401).

Add a check for this special case and avoid return early if this
is the case.

Change-Id: I1a5036d1ea601267459336ecb6ea5a563517e300
Reviewed-on: https://review.couchbase.org/c/query/+/231985
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>