MB-68028 Fix panic when reporting...
 ...change in settings, previously if a
setting's value was a map/slice the comparison
would directly check if they were equal
using Go's equality operator but this would
panic as maps and slices are uncomparable types.

The fix already exists in the trinity and morpheus
branches as part of the commit for MB-51965.
The same changes are backported to the neo branch
in this commit to correctly handle comparsion
of maps/slices when reporting changed settings.

Change-Id: I7789d1a921dbb06ea102b7b42a8094f971ccb129
Reviewed-on: https://review.couchbase.org/c/query/+/232148
Reviewed-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>