MB-69212 Retry mechanism for...
 [BP:MB-67662] ...CREATE SCOPE and CREATE COLLECTION,
specifically to handle the cases where
a bucket is created and a scope is created immediatedly
afterwards, similarly where a scope is created and a
collection is created immediatedly afterwards.

A tracker is added to limit the number of
retries to atmost 32 retries in a minute.

This retry mechanism is only temporary and can be
reverted once the sync ddl feature is implemented.

Change-Id: I8cdad3b7ab60df0e6d0ee4f43e2b78d43b2c8136
Reviewed-on: https://review.couchbase.org/c/query/+/232416
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Gaurav J <gaurav.jayaraj@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/236079
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/236238