MB-68273 Consider hash join when join term has USE KEYS
When using RBO, hash join is only considered when USE HASH hint is
specified on the join term. The only exception to this is when
a SubqueryTerm or ExpressionTerm is used, in which case if a
nested-loop join is used we need to repeatedly evaluate the subquery
or the expression, which can be inefficient.

Similar concern exists as well if the join term has USE KEYS clause,
in which case we fetch the documents using the specified document
keys instead of trying to use an available secondary index. In
this case repeated evaluation of the join term under a nested-loop
join can also be inefficient.

Add logic to allow hash join to be used if the join term has USE
KEYS specified.

This only applied to RBO. In CBO the join is picked by cost.

Change-Id: I479f738bdb965b2ab17617a89dfd1fc925b8ee16
Reviewed-on: https://review.couchbase.org/c/query/+/232981
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>