MB-68305 Proper calculation of index cost for composite vector index
When LIMIT/OFFSET can be pushed down to the index scan, CBO adjusts
the cost of the index scan taking into consideration LIMIT pushdown.
However, this adjustment should not apply to vector index scan,
since in this case the indexer needs to do a topK sort on the vector
distance before returning the results, i.e. it still needs to scan
the entire index spans (no early bail out).

Detect index scans with vector search for composite vector indexes,
and do not apply the normal adjustment for LIMIT pushdown.

This is not an issue for Bhive indexes since the mechanism is
different.

Change-Id: I51faa0172332c17b6fb39dae7be236033442bd6a
Reviewed-on: https://review.couchbase.org/c/query/+/233066
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker