MB-67861: [BP MB 64689] Ensure we always truncate scan results when...
done with them

When an unordered sequential scan was performed with an OFFSET
greater than the number of scan results returned by the first batch of
vb scan requests, the sequential scan operation would hang.

The results returned by the intial set of vb scans, are skipped due
to OFFSET filtering. But these scans need to be re-queued so that
they can get more scan results.

For a scan to be re-queued, the sequential scan sysyem requires
the scan's results to be truncated. However, in this scenario the
scan results were not truncated, the scans were never re-queued and
and the sequential scan operation would not complete.

This caused the hang.

The fix ensures that the scan results are truncated after being
processed

Change-Id: Ic0567507e9c5b01ce8d2bcc294a68648ed412cf1
Reviewed-on: https://review.couchbase.org/c/query/+/233656
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker