MB-68483 Properly determine whether Rerank is requested
The ApproxVectorDistance function has an optional argument (5th argument)
that indicates whether reranking is desired. The issue with this defect
is that in several places in the code when we detect that this argument
is provided we implicitly assume that the argument is specified as true.
In this bug the rerank argument is specified as false, but because we
assume it is true, we assume the reranking is already done by the Bhive
index and thus we omit it from index projection, which leads to incorrect
result (MISSING value used as vector distance).

The fix is to add a member function to ApproxVectorDistance such that
we check for the explicit value (when specified) for rerank. If a query
parameter is ued, the caller dictates whether to treat that as true or
false. We then use the new member function in places as needed.

Change-Id: I411236c9474276dc61daf2246f23747826d98d0c
Reviewed-on: https://review.couchbase.org/c/query/+/233721
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker