MB-68546 Fix more issues with ordered sequential scan
Another issue identified with ordered sequential scan: when larger number
of keys are being retrieved from KV and it reaches the fetch limit, we
remember the last key fetched such that on next round we tell KV where
to continue the scan from. The issue is that when we call
setContinueFromLastKey() we seek to the end of the keys array in order
to get the value of the last key returned. This has the side effect of
setting currentKey to the last key in the array, which subsequently
causes issues.

The fix is to remember where the currentKey is, and after seek() to
the last key and set the value, we seek() again to return it to the
state prior to seeking to last key. This ensures that we don't have
unintended side effects from the function setContinueFromLastKey().

Change-Id: I3fc4029b4169c4cabf06b3a4e4aea87960d47140
Reviewed-on: https://review.couchbase.org/c/query/+/234022
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>