MB-68602 Generate vector index advice in absence of index pushdown
The index advisor currently does not generate an index advice on a query
with vector search if the advised index cannot be used with index pushdown.
For example, if the query does not have a LIMIT clause specified.

We actually generate proper index candidates in this case, but end up
throwing away the index candidates since the vector index candidates are
generated with 'other generator', and the validation phase discards any
index candidates that cannot use index pushdown.

The fix is to mark vector indexes and allow index candidates to be
validated even in absence of index pushdown.

Change-Id: Icf9f49552a3451dcd301c1cda3ec7880727280c5
Reviewed-on: https://review.couchbase.org/c/query/+/234046
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>