MB-68619 Refresh indexer before getting all indexes in UPDATE STATISTICS
When the INDEX option is used in UPDATE STATISTICS, we get the specified
indexes and perform the stats update. The issue is when an index is
created on a different query node, we may not have the most up-to-date
information with regard to available indexes.

The fix is to add a Refresh() call to the indexer before getting all
available indexes.

Change-Id: I809c3a6aaf61ae9d1c3efbb0d55dc9d54a6101a0
Reviewed-on: https://review.couchbase.org/c/query/+/234141
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>