MB-68705  Allow CBO-based advise in absence of memory setting
The CBO-based advise attempts to calculate index statistics for an
advised index. Part of this process involves getting the memory
setting of the cluster in order to calculate a residence ratio.
In our unit test environment, since we use an installed server
with a stand-alone query service, the process of getting a memory
setting does not work properly since the stand-alone query service
uses a dummy ns-server configuration. As a result, we are getting
-1 for memory setting which subsequently causes issues with
costing for the advised index.

The fix is to check that we have a valid memory setting before
attempting to set a residence ratio. In the absence of that,
just use a default. This then allows the CBO-based advise to
work better in our unit test environment.

Also fixed a couple of places where we copy advised index info
to also include setting of CBO-based advise.

Change-Id: I92a68ef48d4588931928e2d6fb55aea1a431bba3
Reviewed-on: https://review.couchbase.org/c/query/+/234374
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker