MB-68718 CBO should account for MIN/MAX optimization in indexer
When aggregate pushdown is done, the indexer can in special cases do
a special optimization for MIN/MAX aggregates:
 - there is only 1 aggregate
 - the index scan only has a single span, or if multiple spans, the
   min/max aggregate is on the leading index key
 - there is no GROUP BY
 - if the aggregate is done on a leading index key, or for a
   non-leading index key, when all preceding index keys have
   equality predicate
 - when the aggregate requested matches index key ordering, i.e.
     for MIN the index key is in ascending order
     for MAX the index key is in descending order
when all the conditions are met, then the indexer can simply return
the first non-missing non-null value found, and terminate the scan.

We now account for this special optimization in CBO costing as well,
by checking the same set of conditions, and once all met, do another
round of costing, effectively adding a "LIMIT 1" to the existing
index costs, and account for potential MISSING/NULL values in case
of MIN aggregate with ascending order (not an issue with MAX
aggregate with descending order).

Change-Id: I8bfc1a089d42c9c1a4521638251caed50969edd5
Reviewed-on: https://review.couchbase.org/c/query/+/234452
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>