MB-68699 Add sanitize function to redact...
 ...values from sql++ query.
A new expression mapper maps all constant
expressions with named parameters in a statement.

The function makes use of this mapper to
sanitize a statement and returns an object
containing the sanitized statement and a map that
associates replaced parameters and the corresponding
original constant value.

If an array or object construct is static,
it is treated as a single value that can be replaced.

If an object is not statically valued then only the
field's value is checked and if static is replaced.
The field's name remains unchanged.

Change-Id: I02c2592108d336eba33d143f542297520305417b
Reviewed-on: https://review.couchbase.org/c/query/+/236663
Well-Formed: Restriction Checker
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>