MB-68768 Allow maximum hash table size to be larger
Currently the maximum hash table size (in terms of number of slots)
is (arbitrarily) fixed at 2^24 (16777216). We have a customer doing
POC that is doing a hash join with more than that number of entries.
To allow such join (will be slow) we add a new bit to n1ql_feat_ctrl
that if set, the maximum hash table size is 4x larger (2^26). Note
currently this new bit only works at cluster-level, if set in
request level it has no effect.

Also changed a few local variable names from "size" to "slots",
since we have another "size" in the hash table that represents the
total number of bytes in the hash table (for memory accounting).

Change-Id: I0c2a40bbc4dfff9e8578d794de6f5f683fba54d7
Reviewed-on: https://review.couchbase.org/c/query/+/234519
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>