MB-69083 Proper handling of aggregate pushdown for FROM clause ...    ... subqueries during join enumeration
This bug is a continuation of MB-65832.

During join enumeration, if we have a FROM clause subquery, and the
subquery can do index group/aggregate pushdown, we delay cover
transformation until join enumeration completes and the final plan
is chosen. Prior to the fix for MB-65832, this cover transformation
(at the end of join enumeration) did not include any pushed down
group/aggregation covers. The fix for MB-65832 now includes the
covers for group keys in case of GROUP pushdown, however, it is
still missing potential transformations for aggregates in case of
partial group pushdown, e.g. where we need to transform a COUNT
into a SUM(cover(COUNT)). This missing transformation causes
potential incorrect results.

The fix builds on top of the fix for MB-65832, and includes:

 - add logic to also perform covering transformations for aggregates
   in group operators (InitialGroup, IntermediateGroup, FinalGroup)
   in situations where the GROUP pushdown is partial and subsequent
   grouping is required after index scan (new function to handle all
   3 group operators)
 - to make that work properly, we also need to exclude the newly
   added group/aggregate covers when performing regular cover
   transformation, such that we don't prematurelly transform the
   aggregate expressions into a cover expression, which then prevent
   it from being properly recognized as an aggregate when group and
   aggregate cover transformation is done separately
 - add new expression flags to mark GROUP/AGGREGATE cover expression
 - add new functions to set the new expression flags when new Cover
   expressions are generated for GROUP/AGGREGATE pushdown
 - add new function to exclude the covers that are marked with the
   new expression flag
 - this exclusion of group/aggregate covers is currently only needed
   when doing cover transformation at the end of join enumeration
 - also added logic when doing aggregate transformation, to avoid
   applying the same logic multiple times (check aggregate flags).

Change-Id: Iee1874cdec69c402f53384649cae07236d5298bf
Reviewed-on: https://review.couchbase.org/c/query/+/235769
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker