MB-65809 Add new parser keyword SAVE for PREPARE statement
Add a new SAVE keyword to PREPARE statement.

The new SAVE keyword behaves similarly as the existing FORCE
keyword, i.e., we don't currently need the actual text for the
token but we use the offset of the token instead.

The offset is used to omit the FORCE and/or SAVE keyword from
the prepared text such that the prepared statement matches.

The new SAVE keyword cannot be added to permitted_identifiers
and thus needs to be escaped if used as a regular identifier.

Change-Id: I7d88371e4b8c85213d046927870616800b059c28
Reviewed-on: https://review.couchbase.org/c/query/+/235945
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>