MB-65809 Auto create QUERY_METADATA bucket when ...    ... SAVE option is used with PREPARE
Add additional handling of the SAVE option for PREPARE, and check and
create (if not exist) the QUERY_METADATA bucket.

The bucket has 100 MB memory quota and currently defaults to magma
bucket with 128 vBuckets (on Linux).

Nothing in the bucket yet. Saved query plans yet to come.

Also updated plan version to 810.

Change-Id: I17f8c579e07fad8ad5479a72d57052f82bdb9bf4
Reviewed-on: https://review.couchbase.org/c/query/+/236531
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>