MB-69387: Backport auto-reprepare feature and fixes related to it
[BP MB 63204]: Provide optional (opt-in) mode to ...

automatically re-prepare prepared statements when indexer metadata
versions change.

Also links sequential scans to the GSI indexer if there is one.
The idea here is that the SequentialScan indexer will report the GSI
indexer's metadata version as its own, so that any GSI index changes
trigger changes for anything using the sequential scan indexer.  The
second part of this is to report the sequential scan index ID as a
string including the metadata version.

This means that when an index is created - and the mode is enabled - a
prepared statement using the sequential scan indexer gets re-prepared.
(It is still free to select the sequential scan of course.)  In the
absence of GSI changes, plans using sequential scans will not be
re-prepared.

Another part of this is to indicate a re-prepare is desired when a
prepared statement's indexer metadata version changes.  The idea for
this part is that we can automatically switch between indexes within an
indexer should a more suitable one be created.

To opt in, the relevant feature control bit must be set in the cluster
configuration.

Also provide a means to manually trigger a re-prepare on next execution.
This is achieved by permitting an UPDATE on system:prepareds - the only
action considered in the update is UNSETting planPreparedTime; all other
actions are reported as errors (unsetting a field that doesn't exist is
a no-op so is silently ignored).

MB-69489: [BP MB 64327] set planPreparedTime in buildPrepare so encoded

plan has planPrepareTime. Add planPreparedTime to marshalInternal,
while buildingEncodedPlan we'll have information about planPreparetime.

Add planPrepareTime in the struct for
unmarshalInternal so distributed plans include
planPrepareTime and are not picked up for (auto)reprepare
unless planPrepareTime is explicitly UNSET.

MB-69490: [BP MB 64237] fixes an issue introduced by MB 64327

Fix empty planPreparedTime field causing unmarshal to fail

Change-Id: I2a159c49577934d1125b778c746366310904ddfe
Reviewed-on: https://review.couchbase.org/c/query/+/236777
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>