MB-69725 : Skip computing values of expressions during...
equivalence checks if either of the expressions being compared is SELF

When an expression in the projection is parsed, we check if the
expression is equivalent to the SELF expression.

But, this equivalence check can evaluate the values of the expressions
to check if they are equivalent. But, SELF's value is always nil and
nil values cannot be meaningfully compared.

This change checks if either of the expressions being compared is
SELF and skips expression value computation if so. This prevents
an unnecessary expression value computation.

Change-Id: I8fe82828efec7c8ccffe19d4e3983010f4e1e07d
Reviewed-on: https://review.couchbase.org/c/query/+/237706
Well-Formed: Restriction Checker
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>