MB-69816 Allow parameter with FieldName to cover
During covering determination, we have a SKIP mechanism such that
if a keyspace reference is NOT the keyspace we are interested in,
we don't need to check for cover for any FieldName references
after the keyspace.

We now extend this mechanism to also apply to situations where
we have a FieldName reference after a query parameter. In this
case a query parameter is similar to a keyspace that is not
interesting.

Also fixed a test case to account for (ignore) extra field for
planPreparedTime in system:prepareds. This part is already fixed
in Morpheus and above.

Change-Id: Ic404782358b8d446395ff69384f72f9cc4479566
Reviewed-on: https://review.couchbase.org/c/query/+/237748
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>