MB-69725: Reduce nesting in Add and Multiplication expressions
When constructing an Add/Multiplication expression, if the first
operand is an Add, "flatten" the structure by extracting their operands
and use these extracted operands directly when constructing the new
expression.

Flattening is applied conservatively to preserve the intentional
grouping (ex. via parantheses) and evaluation order established by
the parser. This is why flattening is applied only to the first input
operand and the operands extracted from it are not recursively
flattened further. And why later input operands are not flattened,
even if they are Add/Multiplication expressions.

Additionally, construct the limit + offset expression using copies of
the limit and offset expressions. This is to prevent any subsequent
changes/transformations made to the limit/offset from affecting
the limit+offset expression

Change-Id: I1f9947cf91cd08c1a6fedeb183de642475c140f7
Reviewed-on: https://review.couchbase.org/c/query/+/238000
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker